Making the feed aspect of this application have its own activity. I will be doing a bit of refactoring to make my life a bit easier on finishing this checkpoint.
