+++ BlocParty/app/src/main/AndroidManifest.xml
@@ -19,6 +19,9 @@
             </intent-filter>
         </activity>
 
+        <activity
+            android:name=".ui.activity.LoginActivity"/>
+
         <meta-data
             android:name="com.facebook.sdk.ApplicationId"
             android:value="@string/fbai" />
