+++ BlocParty/app/src/main/res/layout/activity_login.xml
@@ -1,0 +1,19 @@
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/tb_activity_blocparty_login"
+        android:layout_height="?attr/actionBarSize"
+        android:layout_width="match_parent"
+        android:background="@color/material_indigo_500"/>
+
+    <FrameLayout
+        android:id="@+id/fl_activity_blocparty_login"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/tb_activity_blocparty_login"/>
+
+</RelativeLayout>
